Muninn/requirements.txt
matplotlib
seaborn
discord.py
yt-dlp==2025.6.25
PyNaCl
plexapi
